Update List of biodata students of the 2019 force with as many as 5 people
Message commit : Changes in 2nd and 5th student biodata.
Description commit: Changes in biodata for NIM in 2nd students and changes in biodata for addresses in 5th students.
After completing the biodata data change and filling in the message commit and description commit, go to the commit options section branch. Select options default in the main branch of work or Branch play because for training this time we will focus on learning and understanding how to do commit for file changes at the main branch of work or branch play. If you already feel confident about what you are doing, press the buttonCommit changes to save the results of these changes in repository.